nodeのポイントをuint16_tにした

+++ 04_graph/page_rank/page_rank.hpp
@@ -7,7 +7,7 @@
 
 void updatePageRank(
     std::unordered_map<std::string,
-        std::pair<int,
+        std::pair<uint16_t,
             std::map<std::string, uint8_t>>>& links,
     const std::size_t&& loop_times = 10,
     const uint8_t&& random_rate = 5);
@@ -15,12 +15,12 @@ void updatePageRank(
 void printAllRank(
     const std::unordered_map<std::string, std::string>& pages,
     std::unordered_map<std::string,
-        std::pair<int,
+        std::pair<uint16_t,
             std::map<std::string, uint8_t>>>& links);
 
 void printTopXPages(
     const std::unordered_map<std::string, std::string>& pages,
     std::unordered_map<std::string,
-        std::pair<int,
+        std::pair<uint16_t,
             std::map<std::string, uint8_t>>>& links,
-    int num_of_showing_pages = 10);
+    uint8_t num_of_showing_pages = 10);
